improve memory usage

+++ BSPAT/src/DataType/CpGSite.java
@@ -3,8 +3,8 @@ package DataType;
 public class CpGSite implements CpG {
     private int position;
     private boolean methylLabel;
-    private int methylCount = 0;
-    private int nonMethylCount = 0;
+    private short methylCount = 0;
+    private short nonMethylCount = 0;
 
     public CpGSite(int position, boolean methylLabel) {
         this.methylLabel = methylLabel;
